refactor(provider): extract shared configure helpers; configure no longer writes default
